feat(builder): add editor and preview components for real-time preview (components only)
