+++ components/builder/add-field-dropdown.tsx
@@ -1,0 +1,50 @@
+import { Button } from "@/components/ui/button";
+import {
+	DropdownMenu,
+	DropdownMenuContent,
+	DropdownMenuItem,
+	DropdownMenuTrigger,
+} from "@/components/ui/dropdown-menu";
+import { Plus } from "lucide-react";
+import { getFieldIcon, getFieldTypeName } from "./field-utils";
+
+interface AddFieldDropdownProps {
+	onAddField: (type: string, insertIndex?: number) => void;
+	insertIndex?: number;
+}
+
+export function AddFieldDropdown({
+	onAddField,
+	insertIndex,
+}: AddFieldDropdownProps) {
+	const fieldTypes = [
+		"text",
+		"number",
+		"email",
+		"textarea",
+		"checkbox",
+		"radio",
+	];
+
+	return (
+		<DropdownMenu>
+			<DropdownMenuTrigger asChild>
+				<Button variant="outline" className="gap-2">
+					<Plus className="h-4 w-4" />
+					Add Field
+				</Button>
+			</DropdownMenuTrigger>
+			<DropdownMenuContent>
+				{fieldTypes.map((type) => (
+					<DropdownMenuItem
+						key={type}
+						onClick={() => onAddField(type, insertIndex)}
+					>
+						{getFieldIcon(type)}
+						<span className="ml-2">{getFieldTypeName(type)}</span>
+					</DropdownMenuItem>
+				))}
+			</DropdownMenuContent>
+		</DropdownMenu>
+	);
+}
